+++ NEWS.md
@@ -1,8 +1,9 @@
 # Prova v2.3.0
 
-* Improved plot labels.
 * Introduced the much faster `mutualinfoF()` for variates with finite domains.
 * Introduced the function `eutility()` for the calculation of expected utilities and their revisability, together with `plot()` and `print()` methods.
+* Changed the name of internal S3 classes to avoid clashes with other packages.
+* Improved plot labels.
 * Simplified omission of arguments like `Y =`, `X =`, `K =`.
 
 # Prova v2.1.0