+++ NEWS.md
@@ -1,10 +1,11 @@
 # Prova v2.3.0
 
 * Introduced the much faster `mutualinfoF()` for variates with finite domains.
-* Introduced the function `eutility()` for the calculation of expected utilities and their revisability, together with `plot()` and `print()` methods.
+* Introduced the function `exputility()` for the calculation of expected utilities and their revisability, together with `plot()` and `print()` methods.
 * Changed the name of internal S3 classes to avoid clashes with other packages.
 * Improved plot labels.
 * Updated and corrected documentation.
+* Updated introductory vignette.
 * Simplified omission of arguments like `Y =`, `X =`, `K =`.
 
 # Prova v2.1.0